% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-glm.R
\name{add_rng}
\alias{add_rng}
\title{Attach a random generator function to a family object}
\usage{
add_rng(x, rng = NULL)
}
\arguments{
\item{x}{A family object (as returned by \code{\link[stats:family]{stats::family()}}) or a
list-like object with a \code{family} element.}

\item{rng}{An optional RNG function. If provided, it overrides the
default mapping.}
}
\value{
The same object \code{x}, augmented with an element \code{rng}
containing the chosen random generator function.
}
\description{
Adds (or overrides) a random number generator (RNG) function for
a given model family. By default, common distributions are mapped
to their standard RNG functions:
}
\details{
\itemize{
\item \code{"gaussian"} → \code{\link[stats:Normal]{stats::rnorm()}}
\item \code{"binomial"} → \code{\link[stats:Binomial]{stats::rbinom()}}
\item \code{"poisson"}  → \code{\link[stats:Poisson]{stats::rpois()}}
\item \code{"gamma"}    → \code{\link[stats:GammaDist]{stats::rgamma()}}
}
}
\examples{
f <- stats::binomial()
f <- add_rng(f)
f$rng(5, size = 1, prob = 0.5)  # simulate 5 Bernoulli trials

g <- stats::gaussian()
g <- add_rng(g, rng = function(n) rep(1, n))  # override with custom rng
g$rng(5)

}
