% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{code}
\alias{code}
\title{Format code into rmd/qmd documents}
\usage{
code(x, format = c("html", "markdown", "latex"))
}
\arguments{
\item{x}{text to format}

\item{format}{type of formatting}
}
\value{
the formatted string
}
\description{
Format code into rmd/qmd documents
}
