% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stats.R
\name{cv}
\alias{cv}
\title{Coefficient of Variation}
\usage{
cv(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Logical. Should missing values be removed? Default is \code{FALSE}.}
}
\value{
A single numeric value, the coefficient of variation.
}
\description{
Computes the coefficient of variation (CV) for a numeric vector, defined as
the ratio of the standard deviation to the mean.
}
\details{
The coefficient of variation is defined as
\deqn{CV = \frac{\mathrm{sd}(x)}{\mathrm{mean}(x)}}{}
and provides a normalized measure of dispersion relative to the mean.

If the mean of \code{x} is zero, the result will be \code{Inf}, \code{-Inf}, or \code{NaN}.
}
\examples{
cv(c(1, 2, 3, 4, 5))
cv(c(10, 12, NA, 14), na.rm = TRUE)

}
\seealso{
\code{\link[stats]{sd}}, \code{\link[base]{mean}}
}
