% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dcode}
\alias{dcode}
\title{Show code and results in rmd/qmd documents}
\usage{
dcode(expr, between = NULL, digits = NULL)
}
\arguments{
\item{expr}{code}

\item{between}{string to include between code and result}

\item{digits}{number of digits for numbers}
}
\value{
the string with code and result
}
\description{
Show code and results in rmd/qmd documents. The expression is evaluated in the parent frame.
}
