% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drop}
\alias{drop}
\title{Example dataset of students drop out rates.}
\format{
A data frame with 500 rows and 4 variables:
\describe{
\item{id}{Unique identifier for each student}
\item{parenting}{The parenting style of the student: authoritarian, authoritative, neglectful or permissive}
\item{academic}{Academic success: high or low}
\item{drop}{Whether the student dropped out or not}
}
}
\usage{
drop
}
\description{
This dataset contains data about high-school drop out.
}
\keyword{datasets}
