% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMphd-package.R
\docType{package}
\name{GLMphd-package}
\alias{GLMphd}
\alias{GLMphd-package}
\title{GLMphd: Package for the GLM course at the PhD course in Psychological Sciences, University of Padova}
\description{
Package for the GLM course at the PhD course in Psychological Sciences, University of Padova
}
\author{
\strong{Maintainer}: Filippo Gambarota \email{filippo.gambarota@unipd.it} (\href{https://orcid.org/0000-0002-6666-1747}{ORCID})

}
\keyword{internal}
