% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{latex}
\alias{latex}
\title{Wrapper for the latex2exp::TeX function}
\usage{
latex(text, ...)
}
\arguments{
\item{text}{the string to be passed into latex2exp::TeX}

\item{...}{other objects passed to \code{sprintf(fmt, ...)}}
}
\value{
the formatted expression
}
\description{
Wrapper for the latex2exp::TeX function including also a \code{sprintf} functionality
}
