% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-distr.R
\name{nb.theta2var}
\alias{nb.theta2var}
\title{Variance of a Negative Binomial Distribution}
\usage{
nb.theta2var(mu, theta)
}
\arguments{
\item{mu}{Numeric. The mean(s) of the distribution.}

\item{theta}{Numeric. The dispersion parameter(s).}
}
\value{
A numeric value (or vector) giving the variance.
}
\description{
Computes the variance of a negative binomial distribution given the mean
\eqn{\mu} and the dispersion parameter \eqn{\theta}.
}
\details{
For a negative binomial distribution parameterized by mean \eqn{\mu} and
dispersion \eqn{\theta}, the variance is
\deqn{Var(Y) = \mu + \frac{\mu^2}{\theta}}{}
\itemize{
\item Larger \eqn{\theta} corresponds to less overdispersion (approaching the
Poisson case as \eqn{\theta \to \infty}).
\item Smaller \eqn{\theta} corresponds to greater overdispersion.
}
}
\examples{
nb.theta2var(mu = 5, theta = 2)
nb.theta2var(mu = 10, theta = 100)  # approaches Poisson variance

}
\seealso{
\code{\link{theta_from_vmr}}, \code{\link{vmr_from_theta}}
}
