% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-glm.R
\name{odds}
\alias{odds}
\title{Compute odds from probabilities}
\usage{
odds(p)
}
\arguments{
\item{p}{A numeric vector of probabilities in the range [0, 1].}
}
\value{
A numeric vector of odds corresponding to the input probabilities.
}
\description{
Converts probabilities to odds, using the formula
\deqn{\text{odds}(p) = \frac{p}{1 - p}}
}
\examples{
odds(0.5)   # returns 1
odds(0.8)   # returns 4
odds(c(0.2, 0.5, 0.8))

}
