% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simon}
\alias{simon}
\title{Simon Task}
\format{
A data frame with 25560 rows and 15 variables.
The most important variables in this data set are:
\describe{
\item{id}{A unique identifier for each participant.}
\item{rt}{The reaction time for each trial.}
\item{congruence}{Whether the trial was a congruent or
an incongruent trial.}
\item{acc}{Whether the answer in the current
trial was correct or incorrect.}
\item{trial_type}{Whether the data is from a practice
or a main test trial.}
}
}
\source{
\url{https://raw.githubusercontent.com/michael-franke/intro-data-analysis/master/data_sets/simon-task.csv}

\href{https://michael-franke.github.io/intro-data-analysis/app-93-data-sets-simon-task.html}{The web-book chapter that covers the Simon Task data set}.
}
\usage{
simon
}
\description{
This data set is from an online experiment
in which participants classified a shape presented on the
screen as a square or circle. The shape was displayed on
either the left or right side of the screen.
The dataset is a cleaned and simplified version of https://github.com/michael-franke/aida-package/blob/master/data/data_ST_raw.rda.
}
\keyword{datasets}
